refactor(ca-pi): split the Pi command god module along its three lifecycle seams
commands.ts was 1,033 lines holding three independent subsystems that shared
nothing but the fact that each registers a Pi command: generated alias
expansion with package-ownership validation, the background-job tool and
/ca-jobs controller with its reservation/audit/watcher/teardown state, and the
plan-mode controller with persistence, bridge reads, confirmation, and restore.
Any change to one forced review and re-verification of the other two.

The three now live in command-ownership.ts, native-background.ts, and
native-plan.ts, with the session identity and interactive-parent predicates
both controllers share in session-identity.ts. commands.ts keeps only the
generated command surface (129 lines) plus explicit pass-through re-exports, so
the Pi command surface stays one import boundary for consumers and every
pre-existing suite proves parity unmodified.

Every moved line is byte-identical to its original; the only edited identifier
is the shared CONTROL regex, now exported as CONTROL_CHARACTERS. The moved
crypto lines (SHA-256 audit correlation and plan-content digests, randomUUID
lifecycle seeding) are verbatim moves of approved primitives - no new crypto
surface. The src import graph stays acyclic.

Refs #376
